requirements.txt
Package              Version
-------------------- ---------
absl-py              0.15.0
anyio                3.5.0
asgiref              3.5.1
astor                0.8.1
cached-property      1.5.2
certifi              2020.12.5
charset-normalizer   2.0.12
click                8.1.3
fastapi              0.75.2
gast                 0.2.2
gensim               4.1.2
google-pasta         0.2.0
grpcio               1.42.0
h11                  0.13.0
h5py                 3.6.0
idna                 3.3
importlib-metadata   4.11.3
jieba                0.42.1
Jinja2               3.1.2
joblib               1.0.1
Keras-Applications   1.0.8
Keras-Preprocessing  1.1.2
Markdown             3.3.4
MarkupSafe           2.1.1
mkl-fft              1.3.1
mkl-random           1.2.2
mkl-service          2.4.0
numpy                1.21.5
opt-einsum           3.3.0
pip                  21.2.2
protobuf             3.19.1
pydantic             1.9.0
requests             2.27.1
scikit-learn         0.24.1
scipy                1.7.3
setuptools           61.2.0
six                  1.16.0
sklearn              0.0
smart-open           6.0.0
sniffio              1.2.0
starlette            0.17.1
tensorboard          1.15.0
tensorflow           1.15.0
tensorflow-estimator 1.15.1
termcolor            1.1.0
threadpoolctl        2.1.0
typing_extensions    4.1.1
urllib3              1.26.9
uvicorn              0.17.6
webencodings         0.5.1
Werkzeug             1.0.1
wheel                0.37.1
wrapt                1.13.3
zipp                 3.7.0
